Remove support for extra date/time parsing patterns
